added temperature converter project
In my project, I have built a temperature converter in multiple measurement units.
It will take the type of conversion that the user wants to do and then take the respective temperature and convert it into other measurement units.
When the first conversion is done, then it will ask either the user wants to do another conversion or not.
The user have the option to do more than one conversion and exit anytime they want.